+++ Changelog.md
@@ -5,6 +5,7 @@
 - Increase Malaysia coverage by adding tests for missing Deepavali & Thaipusam.
 - Increase China coverage by adding tests for special extra-holidays & extra-working days cases.
 - Added compatibility with Python 3.8 (#406).
+- Improve ISO registry tests by raising an error when trying to register a non-Calendar class (#412).
 
 ## v7.0.0 (2019-09-20)